fix: align rendered terrain with collision height
Render heightmap terrain with the same floating-point post positions used by the terrain collision cache. The shader mesh was still truncating positions to int to mimic the old glVertex3i display list, which put visible terrain and real shadow receivers up to 0.75 world units below the solver heightfield. That made rolling demo objects look like they were floating above their shadows even while terrain contacts were solving correctly.

Also stop the --shadows command-line override from implicitly enabling cinematic rendering. Shadow maps are supported in normal rendering now, so the CLI override should only control the shadow-map toggle.

Validation:

- tools\validate_full.bat

- Result: VALIDATE_RENDERERS, VALIDATE_PHYSICS, and VALIDATE_PERF all passed; final line was 'VALIDATE_FULL: ALL PHASES PASSED'.

+++ SkullbonezSource/SkullbonezTerrain.cpp
@@ -1004,12 +1004,18 @@ void Terrain::BuildMesh()
             const TerrainPost& p01 = m_postData[idx + m_postsPerSide];
             const TerrainPost& p11 = m_postData[idx + m_postsPerSide + 1];
 
-            // Helper lambda: push m_position (int-truncated) + m_normal + texcoord
+            // Helper lambda: push the same floating-point world position that
+            // collision queries use. The original fixed-function display list
+            // used glVertex3i, but preserving that truncation in the shader mesh
+            // made the rendered terrain sit below the solver heightfield by up
+            // to 0.75 world units. Real shadow maps land on rendered terrain, so
+            // resting balls/boxes looked like they were rolling above their
+            // shadows even though the terrain solver was keeping them supported.
             auto pushVertex = [&]( const TerrainPost& p, float s, float t )
             {
-                vertexData.push_back( static_cast<float>( static_cast<int>( p.vPosition.x ) ) );
-                vertexData.push_back( static_cast<float>( static_cast<int>( p.vPosition.y ) ) );
-                vertexData.push_back( static_cast<float>( static_cast<int>( p.vPosition.z ) ) );
+                vertexData.push_back( p.vPosition.x );
+                vertexData.push_back( p.vPosition.y );
+                vertexData.push_back( p.vPosition.z );
                 vertexData.push_back( p.vNormal.x );
                 vertexData.push_back( p.vNormal.y );
                 vertexData.push_back( p.vNormal.z );
